requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
cytoolz==0.9.0
Django==2.0.1
django-cors-headers==2.1.0
eth-abi==0.5.0
eth-keyfile==0.4.1
eth-keys==0.1.0b4
eth-tester==0.1.0b12
eth-utils==0.7.4
idna==2.6
ipfsapi==0.4.2.post1
py-solc==2.1.0
pycparser==2.18
pycryptodome==3.4.7
pylru==1.0.9
pysha3==1.0.2
pytz==2017.3
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==3.16.4
